requirements.txt
fastmcp
mcp
openai
python-dotenv

